+++ packages/authentication/modules/access/useChangePassword/constants.ts
@@ -2,7 +2,7 @@ import { PASSWORD_REGEX, ZOD_MESSAGE } from '@baseapp-frontend/utils'
 
 import { z } from 'zod'
 
-import type { ChangeExpiredPasswordForm } from './types'
+import type { ChangePasswordForm } from './types'
 
 export const DEFAULT_VALIDATION_SCHEMA = z
   .object({
@@ -17,7 +17,7 @@ export const DEFAULT_VALIDATION_SCHEMA = z
     path: ['confirmNewPassword'],
   })
 
-export const DEFAULT_INITIAL_VALUES: ChangeExpiredPasswordForm = {
+export const DEFAULT_INITIAL_VALUES: ChangePasswordForm = {
   currentPassword: '',
   newPassword: '',
   confirmNewPassword: '',